+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 8.8.3 (2026-02-05)
+
+### Enhancements
+
+- Send managed trace if native trace is empty and update frame.type and error.type setting [#960](https://github.com/bugsnag/bugsnag-unity/pull/960)
+
+### Bug Fixes
+
+- Correctly handle mixed C#/Java stack traces on Android [#962](https://github.com/bugsnag/bugsnag-unity/pull/962)
+
 ## 8.8.2 (2025-12-08)
 
 ### Bug Fixes